+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+### Changes
+
+* Add hardware-inventory collector
+
 ## 1.0.0 / 2020-11-25
 
 ### BREAKING CHANGES